move FrontNewSignalementController to FrontSignalementController #2294
